Improve code, fix bugs
Not transposing all of scratch when propagating in X (only left and right flanks).
Cleaned up code, documented unit tests, now slowly transitioning from 48bin-tests to 32bin-tests (has to be a power of 2, rather than a multiple of 4).
Some unit tests are still failing because I have not fixed them after updating everyOtherInPlace (the important ones are passing still).
Need to move the checks for power of two, and sizes of arrays to initialization.